requirements.txt
django_ckeditor==5.6.1
Django==2.1
django_captcha==1.4.0
